tests update: create new folder "tests" update README.md
