+++ .flake8
@@ -1,2 +1,3 @@
 [flake8]
 max-line-length = 140
+exclude = tests
